Move getJournal context declaration to framework directory
